REFACTOR exit status is not a uint8_t

+++ include/builtins.h
@@ -27,21 +27,21 @@ extern "C" {
 # include "history.h"
 # include <unistd.h>
 
-int	    execute_builtin(t_list *tokenlist, t_data *data);
-int	    execute_cd_builtin(t_darray *env_arr, t_list *tokenlist);
-int     execute_cd_tokenlist(t_darray *env_arr, t_list *tokenlist);
-char    *execute_getcwd(char old_dir[], char *message);
-int	    execute_echo_builtin(t_list *tokenlist);
-int	    execute_env_builtin(t_darray *env_arr, t_list *tokenlist);
+uint8_t		execute_builtin(t_list *tokenlist, t_data *data);
+uint8_t		execute_cd_builtin(t_darray *env_arr, t_list *tokenlist);
+uint8_t		execute_cd_tokenlist(t_darray *env_arr, t_list *tokenlist);
+char    	*execute_getcwd(char old_dir[], char *message);
+uint8_t		execute_echo_builtin(t_list *tokenlist);
+uint8_t		execute_env_builtin(t_darray *env_arr, t_list *tokenlist);
 uint8_t	    execute_exit_builtin(t_data *data, t_list *tokenlist);
-int	    execute_export_builtin(t_darray *env_arr, t_list *tokenlist);
-int	    execute_pwd_builtin(void);
-int     execute_unset_builtin(t_darray *env_arr, t_list *tokenlist);
-int     merge_tokens_for_export(t_list *tokenlist);
-bool	allowed_flags(const char *flag_lexem, const char *allowed_flags);
-bool	write_data(int fd, const void *str, int *status);
-bool	ft_isnumstring(const char *str);
-bool    read_only_variable(const char *key);
+uint8_t		execute_export_builtin(t_darray *env_arr, t_list *tokenlist);
+uint8_t		execute_pwd_builtin(void);
+uint8_t		execute_unset_builtin(t_darray *env_arr, t_list *tokenlist);
+int    		merge_tokens_for_export(t_list *tokenlist);
+bool		allowed_flags(const char *flag_lexem, const char *allowed_flags);
+bool		write_data(int fd, const void *str, uint8_t *status);
+bool		ft_isnumstring(const char *str);
+bool   		read_only_variable(const char *key);
 
 #  ifdef __cplusplus
